requirements.txt
channels==3.0.5
channels-redis==3.4.1
Django==4.0.6
django-cors-headers==3.13.0
django-redis==5.2.0
djangorestframework==3.13.1
asgiref~=3.5.2
celery~=5.2.7